requirements.txt
#admin-multiupload==1.10
#git+https://github.com/gkuhn1/django-admin-multiupload.git
#git+https://github.com/python-force/django-admin-multiupload.git
Django==2.0.2
#django-blog-zinnia==0.19
git+https://github.com/Wtower/django-blog-zinnia.git
django-ckeditor==5.4.0
django-constance==2.1.0
django-debug-toolbar==1.9.1
django-imagekit==4.0.2
django-jet==1.0.7
#django-modeltranslation==0.12.1
git+https://github.com/deschler/django-modeltranslation.git
django-picklefield==1.0.0
django-rosetta==0.8.1
feedparser==5.2.1
google-api-python-client==1.4.1
mock==2.0.0
python-memcached==1.59
pytz==2018.3
Unipath==1.1
selenium
django-celery-beat
websockets
requests
lxml
#mysql-connector-python
#mysqlclient
psycopg2
psycopg2-binary
redis
whitenoise
